move /scratch/new to /new
