+++ frontend/src/pages/new.module.scss
