+++ frontend/src/pages/new.tsx
@@ -4,17 +4,17 @@ import { GetStaticProps } from "next"
 
 import { useRouter } from "next/router"
 
-import AsyncButton from "../../components/AsyncButton"
-import { CompilerPreset } from "../../components/compiler/CompilerOpts"
-import { useCompilersForPlatform } from "../../components/compiler/compilers"
-import PresetSelect, { PRESETS } from "../../components/compiler/PresetSelect"
-import Editor from "../../components/Editor"
-import Footer from "../../components/Footer"
-import Nav from "../../components/Nav"
-import PageTitle from "../../components/PageTitle"
-import PlatformSelect from "../../components/PlatformSelect"
-import Select from "../../components/Select2"
-import * as api from "../../lib/api"
+import AsyncButton from "../components/AsyncButton"
+import { CompilerPreset } from "../components/compiler/CompilerOpts"
+import { useCompilersForPlatform } from "../components/compiler/compilers"
+import PresetSelect, { PRESETS } from "../components/compiler/PresetSelect"
+import Editor from "../components/Editor"
+import Footer from "../components/Footer"
+import Nav from "../components/Nav"
+import PageTitle from "../components/PageTitle"
+import PlatformSelect from "../components/PlatformSelect"
+import Select from "../components/Select2"
+import * as api from "../lib/api"
 
 import styles from "./new.module.scss"
